+++ lib/abuseipdb/responses/report.rb
@@ -1,0 +1,7 @@
+module Abuseipdb
+  module Responses
+    class Report < Base
+
+    end
+  end
+end