+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+### Added
+- Report Endpoint
+
 ## [0.0.1] - 2019-06-12
 ### Added
 - Check Endpoint